Fix max current for rear hub
Fix max current parameter for rear hub motor. While the peak current is
configured to be 24.0 A in the Copley motor drive settings, the
max_current parameter does not distinguish between peak and continuous
current settings. In order to prevent faulting of the motor controller,
this value must be set to the allowable continous current value.